shopping cart moved under 'pages' folder
