+++ ecommerce/frontend/src/components/pages/ShoppingCartPage.less
