+++ ecommerce/frontend/src/mocks/handlers.js
@@ -14,66 +14,60 @@ const url = u => config.apiUrl + u
 
 export const handlers = [
     // url(...) is important here !!
-    rest.get(url('/example/user/:userId'), (req, res, ctx) => {
-        const { params, body } = req
-        const { userId } = params
-
-        if (!userId) return res(ctx.status(403), ctx.json({ successful: false, message: `Bad request` }))
-
-        return res(ctx.json({ successful: true, user: { id: userId, name: 'Ali', surname: 'Batır' } }))
-    }),
-    rest.get(url('/products/homepage/:count'), (req, res, ctx) => {
-        const { params, body } = req
-        const { count } = params
-
-        return res(ctx.json({ successful: true, products: trendingProducts.slice(0, count) }))
-    }),
-    rest.get(url('/product/:productId'), (req, res, ctx) => {
-        const { params, body } = req
-        const { productId } = params
-
-        return res(ctx.json(product))
-    }),
-    rest.post(url('/search/products'), (req, res, ctx) => {
-        const { params, body } = req
-        const { page_size = 10, page = 0 } = body
-
-        return res(
-            ctx.json({
-                data: {
-                    pagination: {
-                        page_size,
-                        page,
-                        total_items: products.length,
-                    },
-                    products: products.slice(page * page_size, (page + 1) * page_size),
-                },
-            }),
-        )
-    }),
+    // rest.get(url('/example/user/:userId'), (req, res, ctx) => {
+    //     const { params, body } = req
+    //     const { userId } = params
+    //     if (!userId) return res(ctx.status(403), ctx.json({ successful: false, message: `Bad request` }))
+    //     return res(ctx.json({ successful: true, user: { id: userId, name: 'Ali', surname: 'Batır' } }))
+    // }),
+    // rest.get(url('/products/homepage/:count'), (req, res, ctx) => {
+    //     const { params, body } = req
+    //     const { count } = params
+    //     return res(ctx.json({ successful: true, products: trendingProducts.slice(0, count) }))
+    // }),
+    // rest.get(url('/product/:productId'), (req, res, ctx) => {
+    //     const { params, body } = req
+    //     const { productId } = params
+    //     return res(ctx.json(product))
+    // }),
+    // rest.post(url('/search/products'), (req, res, ctx) => {
+    //     const { params, body } = req
+    //     const { page_size = 10, page = 0 } = body
+    //     return res(
+    //         ctx.json({
+    //             data: {
+    //                 pagination: {
+    //                     page_size,
+    //                     page,
+    //                     total_items: products.length,
+    //                 },
+    //                 products: products.slice(page * page_size, (page + 1) * page_size),
+    //             },
+    //         }),
+    //     )
+    // }),
     // rest.get(url('/email-verify/:token'), (req, res, ctx) => {
     //     const { params, body } = req
     //     const { token } = params
-
     //     return res(ctx.json({ data: { message: verifications[Math.floor(Math.random() * verifications.length)] } }))
     // }),
-    rest.get(url('/review'), (req, res, ctx) => {
-        const id = req.url.searchParams.get('product')
-        const page_size = req.url.searchParams.get('page_size')
-        const page = req.url.searchParams.get('page')
-        return res(
-            ctx.json({
-                data: {
-                    pagination: {
-                        page_size,
-                        page,
-                        total_items: reviews.length,
-                    },
-                    reviews: reviews.slice(page * page_size, (page + 1) * page_size),
-                },
-            }),
-        )
-    }),
+    // rest.get(url('/review'), (req, res, ctx) => {
+    //     const id = req.url.searchParams.get('product')
+    //     const page_size = req.url.searchParams.get('page_size')
+    //     const page = req.url.searchParams.get('page')
+    //     return res(
+    //         ctx.json({
+    //             data: {
+    //                 pagination: {
+    //                     page_size,
+    //                     page,
+    //                     total_items: reviews.length,
+    //                 },
+    //                 reviews: reviews.slice(page * page_size, (page + 1) * page_size),
+    //             },
+    //         }),
+    //     )
+    // }),
     // rest.get(url('/review'), (req, res, ctx) => {
     //     const id = req.url.searchParams.get('product')
     //     const page_size = req.url.searchParams.get('page_size')
